feat(favorites): add favorites page with suspense loading
Extract favorites content into separate component and implement suspense loading with skeleton UI. This improves code organization and provides better user experience during loading.
