+++ app/dashboard/favorites/page.tsx
@@ -1,9 +1,10 @@
 import { Metadata } from "next";
-import DashboardLayout from "../../../components/dashboard-layout";
-import FileGrid from "../../../components/file-grid";
-import { getEmailFromUserId } from "../../../lib/getEmailFromUserId";
-import { dbAdmin } from "../../../lib/firebase-admin";
+import { Suspense } from "react";
+
 import { headers } from "next/headers";
+import DashboardLayout from "../../../components/dashboard-layout";
+import FileSkeleton from "../../../components/FileSkeleton";
+import FavoriteContent from "../../../components/pages/favorites";
 
 export const metadata: Metadata = {
   title: "StellarSync | Favorites",
@@ -11,7 +12,7 @@ export const metadata: Metadata = {
 
 export const dynamic = "force-dynamic";
 
-export default async function FavoritesPage() {
+export default async function DashboardPage() {
   await headers();
   const headersList = await headers();
   const userId = headersList.get("x-user-id");
@@ -19,7 +20,6 @@ export default async function FavoritesPage() {
   if (!userId) {
     return (
       <DashboardLayout>
-        <p className="p-6"></p>
         <div className="col-span-full flex h-40 flex-col items-center justify-center rounded-lg border border-dashed p-4 text-center">
           <p className="text-lg font-medium">Not signed in.</p>
         </div>
@@ -27,31 +27,11 @@ export default async function FavoritesPage() {
     );
   }
 
-  const email = await getEmailFromUserId(userId);
-
-  let files: Array<any> = [];
-  const base = dbAdmin
-    .collection("uploadedFiles")
-    .where("userEmail", "==", email);
-
-  const snap = await base.get();
-  snap.docs.forEach((d) => files.push({ id: d.id, ...d.data() }));
-
   return (
     <DashboardLayout>
-      <div className="flex flex-col gap-6">
-        <div className="flex flex-col gap-2">
-          <h1 className="text-2xl font-bold tracking-tight">Favorites</h1>
-          <p className="text-muted-foreground">
-            Quick access to your bookmarked files.
-          </p>
-        </div>
-
-        <FileGrid
-          fileList={files.filter((file) => file.stared)}
-          view="favorites"
-        />
-      </div>
+      <Suspense fallback={<FileSkeleton />}>
+        <FavoriteContent userId={userId} />
+      </Suspense>
     </DashboardLayout>
   );
 }
